wochenTest und aufgaben Controller hinzugefügt, aber noch aufgabe controller hat problem

+++ src/main/java/com/example/exambyte/applicationService/serviceRepository/AufgabenRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface AufgabenRepository {
     Aufgabe save(Aufgabe aufgabe);
-    Aufgabe findById(Long id);
+    Aufgabe findById(int id);
     List<Aufgabe> findAll();
     void delete(Aufgabe aufgabe);
 }
